Rename simple/example* to simple/simple*
