+++ examples/tests/simple/simple_test.go
@@ -1,4 +1,4 @@
-package example
+package simple
 
 import (
 	"testing"
